Moved sipXconfig related changes from https://scm.sipfoundry.org/rep/*/main repositories: - 1068:1112 from  https://scm.sipfoundry.org/rep/sipXconfig/main - 596:598 from https://scm.sipfoundry.org/rep/sipXportLib/main - 578:579 from https://scm.sipfoundry.org/rep/sipXpbx/main
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@4835 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXconfig/web/src/org/sipfoundry/sipxconfig/site/phone/AddPhoneUser.java
@@ -63,7 +63,7 @@ public abstract class AddPhoneUser extends BasePage implements PageRenderListene
         Iterator usersIds = selections.getAllSelected().iterator();
         while (usersIds.hasNext()) {
             Integer userId = (Integer) usersIds.next();
-            User user = getCoreContext().loadUser(userId.intValue());            
+            User user = getCoreContext().loadUser(userId);            
             Line line = phone.createLine(new LineData());            
             line.getLineData().setUser(user);
             phone.addLine(line);
